--- src/Pages/HomePage.jsx
@@ -1,15 +1,0 @@
-import HomeFooter from "../Components/AdminDashboard/home/HomeFooter";
-import HomeLayout from "../Components/AdminDashboard/home/HomeLayout";
-import HomeNav from "../Components/AdminDashboard/home/HomeNav";
-
-function Home() {
-  return (
-    <div className="w-[95%] flex flex-col gap-4 mx-auto">
-      <HomeNav />
-      <HomeLayout />
-      <HomeFooter />
-    </div>
-  );
-}
-
-export default Home;
